requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
configparser==3.5.0
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gnuplotlib==0.20
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.2.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.4.1
numpy==1.14.2
numpysane==0.13
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
Pympler==0.5
pyparsing==2.2.0
pyrouge==0.1.3
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scandir==1.7
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
tqdm==4.19.8
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.0
